New/No OP Reply system
This commit adds the No OP Replies system.

All submissions are marked as new and the number of OP posts are
recorded (in a hidden CSS class).

This new tag lasts for a set amount of time after which it will be
decided whether the post has enough replies or not.

It also checks submissions marked as No Replies for updates.